Begin integration of Auth Cookie
It seems what I have so far *should* work but I dont really know what every line of code is truly doing.
Need to do more research and work.
But I THINK we are at least half-way there.

Unfortunate side-effect is that you can no longer log in with swagger-ui. It is expecting the login function to return the token and NOT return a `Response` with a cookie. ::shrugs:: oh well.

Anway, great and solid start! YES! lol
Now, relax and sleep time.