tjr214/cookies: Continue integration of Auth Cookie
Everything mostly works and I now understand what is going on under the hood.
To continue forward (and to test out create user, login user, etc etc) it seems I need to dedicate some time to learning frontend.
That is cool, lol.
But it also means features like Refresh Tokens and Github OAuth integration will be stalled for a bit.
BUT it ALSO means that I will have a super small mini-front end up and running soon!
WAHOO!

But for now: rest and food.
PEACE.