htmx-frontend-test: Add login page
Ha-ha! With the add of Flowbite, Figma, and Claude.ai -- I finally have a simple login page up and running!
It doesnt *DO* anything yet, lol, but at least it is there. I mean, wowow, I hate coding frontends haha
Next step is to wire up the Login and Signin with Github buttons with the code we already have for login.

And, oh boy, that is probably a whole new can of worms being opened!!
OK, time to go now.
PEACE.