htmx-frontend-test: Get login page/login working
Hey! The login shit work!
WaHOOOO!

We still dont have an actual user homepage once the user is logged in.
BUT: the login itself works, and sets the right cookie to hold the Token, AND
(with new `get_user_from_cookie()` function) we can now validate the user via
this cookie! AND IT WORKS! Oh yes! So much time spent on this -- and now it works.
That makes me happy :)

Also: when the Token or cookie becomes invalid, we are correctly redirected to the login page.
Github OAuth is also working (up to the point that it was before -- still need to integrate it
with "our" user stuff).

Every day we work on this is one step closer to The Goal!
fr fr

PEACE.